API endpoint works? Or maybe it's incompatible...
